Applied requested changes
