+++ frontend/src/routes/index.svelte
@@ -6,5 +6,5 @@
 	<a href="/track-attendance" rel="prefetch">Track attendance</a>
 	<a href="/check-hours" rel="prefetch">View attendance statistics</a>
 	
-	<a href="/choose-user" rel="prefetch">Choose user</a>
+	<a href="/login" rel="prefetch">Choose user</a>
 </div>
